moved validatePlugin to a separate module (Haskell.Ide.Engine.Utils). Perform validation in main, before the dispatcher runs
